+++ pip_audit/_service/__init__.py
@@ -2,6 +2,7 @@
 Vulnerability service interfaces and implementations for `pip-audit`.
 """
 
+from .esms import EcosystemsService
 from .interface import (
     ConnectionError,
     Dependency,
@@ -15,6 +16,7 @@ from .osv import OsvService
 from .pypi import PyPIService
 
 __all__ = [
+    "EcosystemsService",
     "ConnectionError",
     "Dependency",
     "ResolvedDependency",